+++ dcli/src/formatters/mod.rs
@@ -11,11 +11,12 @@ pub(crate) mod formatter;
 pub(crate) mod list_table;
 pub(crate) mod manifest;
 pub(crate) mod proxy;
-pub(crate) mod show_table;
 pub(crate) mod settings;
+pub(crate) mod show_table;
 #[cfg(feature = "stream")]
 pub(crate) mod stream;
 pub(crate) mod string_table;
+pub(crate) mod target;
 pub(crate) mod task;
 pub(crate) mod topic;
 pub(crate) mod usage;