+++ dcli/src/formatters/settings.rs
@@ -1,19 +1,23 @@
 use crate::formatters::formatter::{Label, SubjectFormatter};
-use crate::settings::Target;
+use crate::settings::Settings;
 
 #[derive(Eq, Hash, PartialEq)]
-pub(crate) enum TargetLabel {
-  Platform,
-  Tenant,
-  GroupUserId,
+pub(crate) enum SettingsLabel {
+  DefaultPlatform,
+  DefaultTenant,
+  ShowExecutionTime,
+  NoBorder,
+  Verbosity,
 }
 
-impl Label for TargetLabel {
+impl Label for SettingsLabel {
   fn label_for_show(&self) -> &str {
     match self {
-      Self::GroupUserId => "id",
-      Self::Platform => "platform",
-      Self::Tenant => "tenant",
+      SettingsLabel::DefaultPlatform => "default-platform",
+      SettingsLabel::DefaultTenant => "default-tenant",
+      SettingsLabel::ShowExecutionTime => "show-execution-time",
+      SettingsLabel::NoBorder => "no-border",
+      SettingsLabel::Verbosity => "verbosity",
     }
   }
 
@@ -22,18 +26,37 @@ impl Label for TargetLabel {
   }
 }
 
-impl SubjectFormatter<TargetLabel> for Target {
-  fn value(&self, label: &TargetLabel, _target_id: &str) -> String {
+impl SubjectFormatter<SettingsLabel> for Settings {
+  fn value(&self, label: &SettingsLabel, _target_id: &str) -> String {
     match label {
-      TargetLabel::GroupUserId => self.group_user_id.clone(),
-      TargetLabel::Platform => self.platform.to_string(),
-      TargetLabel::Tenant => self.tenant.clone()
+      SettingsLabel::DefaultPlatform => self
+        .default_platform
+        .clone()
+        .unwrap_or("not set, provide platform as a command line argument, else the user will be promted".to_string()),
+      SettingsLabel::DefaultTenant => self
+        .default_tenant
+        .clone()
+        .unwrap_or("not set, provide platform as a command line argument, else the user will be promted".to_string()),
+      SettingsLabel::ShowExecutionTime => self
+        .show_execution_time
+        .map(|show_execution_time| show_execution_time.to_string())
+        .unwrap_or("not set, default depends on 'verbosity' setting".to_string()),
+      SettingsLabel::NoBorder => self
+        .no_border
+        .map(|no_border| no_border.to_string())
+        .unwrap_or("not set, default is to show borders".to_string()),
+      SettingsLabel::Verbosity => self
+        .verbosity
+        .clone()
+        .map(|verbosity| verbosity.to_string())
+        .unwrap_or("not set, defaults to 'low'".to_string()),
     }
   }
 
-  fn target_label(&self) -> Option<TargetLabel> {
+  fn target_label(&self) -> Option<SettingsLabel> {
     None
   }
 }
 
-pub static TARGET_LABELS: [TargetLabel; 3] = [TargetLabel::Platform, TargetLabel::Tenant, TargetLabel::GroupUserId];
+pub static SETTING_LABELS: [SettingsLabel; 5] =
+  [SettingsLabel::DefaultPlatform, SettingsLabel::DefaultTenant, SettingsLabel::ShowExecutionTime, SettingsLabel::NoBorder, SettingsLabel::Verbosity];